proyectos/p01-inteligencia-comercial-internacional/requirements.txt
groq==0.13.0
tavily-python==0.5.0
streamlit==1.44.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.11.3
python-docx==1.1.2
pandas==2.2.2
pytest==8.3.2